Added sorting code and a Node class
Update 1: I added some very useful methods to shuffle and sort 2D arrays. These codes have a built in comparator and pre-shuffle the array, to avoid Quicksort TLEs.

Update 2: I added my experimental Node class, which uses a HashSet to allow for graphs to be quickly modified and queried. It is quite slow though, so only use it when removing nodes is key.