NEW - Heavy Light Decomposition Template
I uploaded my fully functional HLD template. Simply input the edges of the tree and the initial values of the nodea, and specify the merge operation (usually sum). Then, whenever you want to update a node simply call update, and when you want to query all the nodes from the simple path from a to b, simply use the query function.

The time complexity of HLD is O(N(logN)^2)